// registry-server/prisma/generated/zod/inputTypeSchemas/FurnitureWhereInputSchema.ts
import type { Prisma } from '../../../../lib/generated/prisma';

import { z } from 'zod';
import { StringFilterSchema } from './StringFilterSchema';
import { IntFilterSchema } from './IntFilterSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { RoomScalarRelationFilterSchema } from './RoomScalarRelationFilterSchema';
import { RoomWhereInputSchema } from './RoomWhereInputSchema';

export const FurnitureWhereInputSchema: z.ZodType<Prisma.FurnitureWhereInput> = z.object({
  AND: z.union([ z.lazy(() => FurnitureWhereInputSchema),z.lazy(() => FurnitureWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => FurnitureWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => FurnitureWhereInputSchema),z.lazy(() => FurnitureWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  label: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  x: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  y: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  z: z.union([ z.lazy(() => IntFilterSchema),z.number() ]).optional(),
  roomId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  Room: z.union([ z.lazy(() => RoomScalarRelationFilterSchema),z.lazy(() => RoomWhereInputSchema) ]).optional(),
}).strict() as z.ZodType<Prisma.FurnitureWhereInput>;

export default FurnitureWhereInputSchema;